HHH-11147 Call the interceptor of the current instance
In the case of a hierarchy, the parent class would use the interceptor
injected into the child class. The interceptor would then be called
several times when the hierarchy of getter/setter is visited.

We now use the interceptor of the current level of hierarchy: the child
class uses the one injected, the parent class doesn't have an
interceptor injected in this case so wouldn't intercept the calls.

Per suggestion from Gail.